requirements.txt
matplotlib
numpy>=1.26.0
pyyaml
rawpy
scipy
tqdm
rawpy
tifffile
cython
imagecodecs
#opencv-contrib-python
#opencv-python
numba
# Optional: Halide for GPU acceleration (install manually if needed)
# halide
# cupy 
